move functions from pedigree_utils to avoid circular dependencies
